dispatch perf example in several files + new namespaces: configuration, cas_cible_1 and fissuration + rewrite compilation step with libraries (CMakelist) + add petsc features but these features have not been tested
