+++ ex6/CMakeLists.txt
@@ -50,78 +50,129 @@ endif()
 
 
 if(MFEM_USE_MPI)
-  # Define the executable to build for this example
-  add_executable(perf perf.cxx)
-  target_link_libraries(perf mfem-mgis::MFEMMGIS)
-  target_link_libraries(perf timers)
 
-  target_include_directories(perf
-      SYSTEM
-      PUBLIC "$<BUILD_INTERFACE:${MFEM_INCLUDE_DIRS}>"
-      PUBLIC "$<INSTALL_INTERFACE:${MFEM_INCLUDE_DIRS}>"
-      )
+	SET(utiles test_parameters.cpp solver_name.cxx precond_name.cxx data_gathering.cxx)
+	add_library(configuration STATIC ${utiles})
+	target_link_libraries(configuration mfem-mgis::MFEMMGIS)
+	target_include_directories(configuration
+		SYSTEM  
+		PUBLIC "$<BUILD_INTERFACE:${MFEM_INCLUDE_DIRS}>"
+		PUBLIC "$<INSTALL_INTERFACE:${MFEM_INCLUDE_DIRS}>"
+		PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}"
+		)
 
-  add_executable(reader read_markdown.cxx)
-  target_link_libraries(reader mfem-mgis::MFEMMGIS)
-  target_link_libraries(reader timers)
 
-  target_include_directories(reader
-      SYSTEM
-      PUBLIC "$<BUILD_INTERFACE:${MFEM_INCLUDE_DIRS}>"
-      PUBLIC "$<INSTALL_INTERFACE:${MFEM_INCLUDE_DIRS}>"
-      )
-  
-  add_executable(one_test one_test.cxx)
-  target_link_libraries(one_test mfem-mgis::MFEMMGIS)
-  target_link_libraries(one_test timers)
-  
-  target_include_directories(one_test
-      SYSTEM
-      PUBLIC "$<BUILD_INTERFACE:${MFEM_INCLUDE_DIRS}>"
-      PUBLIC "$<INSTALL_INTERFACE:${MFEM_INCLUDE_DIRS}>"
-      )
+	function(add_case exe name)
+		target_link_libraries(${exe} ${name})
+	endfunction()
 
-  # Copy needed file for execution
-  file(COPY
-    ${PROJECT_SOURCE_DIR}/Elasticity.mfront
-    DESTINATION ${PROJECT_BINARY_DIR})
+	function(build_case name)
+		add_library(${name} STATIC ${name}.cxx)
+		target_link_libraries(${name} mfem-mgis::MFEMMGIS)
+		target_include_directories(${name}
+			SYSTEM  
+			PUBLIC "$<BUILD_INTERFACE:${MFEM_INCLUDE_DIRS}>"
+			PUBLIC "$<INSTALL_INTERFACE:${MFEM_INCLUDE_DIRS}>"
+			PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}"
+			)
+	endfunction()
 
-  ## mesh
-  file(COPY
-	${PROJECT_SOURCE_DIR}/cube_2mat_per.mesh
-    	DESTINATION ${PROJECT_BINARY_DIR})
+	build_case(cas_cible_1)
+	build_case(fissuration)
 
-  # Use mfront to generate libBehaviour.so
-  add_custom_command(
-    OUTPUT "src/libBehaviour.so"
-    COMMAND mfront --obuild --interface=generic ${PROJECT_BINARY_DIR}/Elasticity.mfront
-    DEPENDS ${PROJECT_BINARY_DIR}/Elasticity.mfront
-    COMMENT "Performing mfront code generation")
-  
-  add_custom_target(generate-libBehaviour-cas-cible-1 ALL
-    DEPENDS "src/libBehaviour.so")
-  
-  add_dependencies(check perf generate-libBehaviour-cas-cible-1)
-  
-  # Define a test that will be called during the "make check"
-  function(add_perf)
-	  set(test "perf")
-    add_test(NAME ${test}_cube_2mat_per COMMAND mpirun -n 3 ./perf --mesh cube_2mat_per.mesh) #see later why it fails with -n 4 and it works with n = 1,2,3,5,6,7,44
-  endfunction(add_perf)
 
-  function(add_solver_preconditioner)
-	  #no foreach
-    add_test(NAME periodic_hypreCG_HypreDiagScale_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 0 --mesh cube_2mat_per.mesh) 
-    add_test(NAME periodic_hypreCG_HypreBoomerAMG_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 3 --mesh cube_2mat_per.mesh) 
-    add_test(NAME periodic_hypreCG_HypreILU_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 4 --mesh cube_2mat_per.mesh) 
-    add_test(NAME periodic_hypreCG_any_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 5 --mesh cube_2mat_per.mesh) 
-    add_test(NAME periodic_CGSolver_HypreDiagScale_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 6 --preconditioner 0 --mesh cube_2mat_per.mesh) 
-    add_test(NAME periodic_CGSolver_HypreParaSails_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 6 --preconditioner 1 --mesh cube_2mat_per.mesh) 
-    add_test(NAME periodic_CGSolver_any_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 6 --preconditioner 5 --mesh cube_2mat_per.mesh) 
-  endfunction(add_solver_preconditioner)
+	# Define the executable to build for this example
+	add_executable(perf perf.cxx)
+	target_link_libraries(perf mfem-mgis::MFEMMGIS)
+	target_link_libraries(perf timers)
+	target_link_libraries(perf configuration)
+	add_case(perf cas_cible_1)
+	add_case(perf fissuration)
 
-  add_perf()
-  add_solver_preconditioner()
+	target_include_directories(perf
+		SYSTEM
+		PUBLIC "$<BUILD_INTERFACE:${MFEM_INCLUDE_DIRS}>"
+		PUBLIC "$<INSTALL_INTERFACE:${MFEM_INCLUDE_DIRS}>"
+		)
+
+
+	add_executable(reader read_markdown.cxx)
+	target_link_libraries(reader mfem-mgis::MFEMMGIS)
+	target_link_libraries(reader timers)
+	target_link_libraries(reader configuration)
+
+	target_include_directories(reader
+		SYSTEM
+		PUBLIC "$<BUILD_INTERFACE:${MFEM_INCLUDE_DIRS}>"
+		PUBLIC "$<INSTALL_INTERFACE:${MFEM_INCLUDE_DIRS}>"
+		)
+
+	add_executable(one_test one_test.cxx)
+	target_link_libraries(one_test mfem-mgis::MFEMMGIS)
+	target_link_libraries(one_test timers)
+	target_link_libraries(one_test configuration)
+
+	target_include_directories(one_test
+		SYSTEM
+		PUBLIC "$<BUILD_INTERFACE:${MFEM_INCLUDE_DIRS}>"
+		PUBLIC "$<INSTALL_INTERFACE:${MFEM_INCLUDE_DIRS}>"
+		)
+
+	add_case(one_test cas_cible_1)
+
+	add_executable(petsc_file_generation use_petsc.cxx)
+	target_link_libraries(petsc_file_generation mfem-mgis::MFEMMGIS)
+	target_link_libraries(petsc_file_generation timers)
+	target_link_libraries(petsc_file_generation configuration)
+
+	target_include_directories(petsc_file_generation
+		SYSTEM
+		PUBLIC "$<BUILD_INTERFACE:${MFEM_INCLUDE_DIRS}>"
+		PUBLIC "$<INSTALL_INTERFACE:${MFEM_INCLUDE_DIRS}>"
+		)
+
+
+	# Copy needed file for execution
+	file(COPY
+		${PROJECT_SOURCE_DIR}/Elasticity.mfront
+		DESTINATION ${PROJECT_BINARY_DIR})
+
+	## mesh
+	file(COPY
+		${PROJECT_SOURCE_DIR}/cube_2mat_per.mesh
+		DESTINATION ${PROJECT_BINARY_DIR})
+
+	# Use mfront to generate libBehaviour.so
+	add_custom_command(
+		OUTPUT "src/libBehaviour.so"
+		COMMAND mfront --obuild --interface=generic ${PROJECT_BINARY_DIR}/Elasticity.mfront
+		DEPENDS ${PROJECT_BINARY_DIR}/Elasticity.mfront
+		COMMENT "Performing mfront code generation")
+
+	add_custom_target(generate-libBehaviour-cas-cible-1 ALL
+		DEPENDS "src/libBehaviour.so")
+
+	add_dependencies(check perf generate-libBehaviour-cas-cible-1)
+
+	# Define a test that will be called during the "make check"
+	function(add_perf)
+		set(test "perf")
+		add_test(NAME ${test}_cube_2mat_per COMMAND mpirun -n 3 ./perf --mesh cube_2mat_per.mesh) #see later why it fails with -n 4 and it works with n = 1,2,3,5,6,7,44
+	endfunction(add_perf)
+
+	function(add_solver_preconditioner)
+		#no foreach
+		add_test(NAME periodic_hypreCG_HypreDiagScale_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 0 --mesh cube_2mat_per.mesh) 
+		add_test(NAME periodic_hypreCG_HypreBoomerAMG_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 3 --mesh cube_2mat_per.mesh) 
+		add_test(NAME periodic_hypreCG_HypreILU_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 4 --mesh cube_2mat_per.mesh) 
+		add_test(NAME periodic_hypreCG_any_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 3 --preconditioner 5 --mesh cube_2mat_per.mesh) 
+		add_test(NAME periodic_CGSolver_HypreDiagScale_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 6 --preconditioner 0 --mesh cube_2mat_per.mesh) 
+		add_test(NAME periodic_CGSolver_HypreParaSails_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 6 --preconditioner 1 --mesh cube_2mat_per.mesh) 
+		add_test(NAME periodic_CGSolver_any_cube_2mat_per COMMAND mpirun -n 3 ./one_test -ls 6 --preconditioner 5 --mesh cube_2mat_per.mesh) 
+	endfunction(add_solver_preconditioner)
+
+	add_perf()
+	add_solver_preconditioner()
 
 endif()
