+++ ex6/one_test.cxx
@@ -45,75 +45,57 @@
 #include "precond_name.hxx"
 #include "config_solver.hxx"
 #include "data_gathering.hxx"
-
-
-constexpr double xmax = 1.;
-
-
-struct TestParameters {
-	const char* mesh_file = "cas-cible-1.mesh";
-	const char* behaviour = "Elasticity";
-	const char* library = "src/libBehaviour.so";
-	const char* reference_file = "Elasticity.ref";
-	int order = 1;
-	int tcase = 1;
-	int linearsolver = -1;
-	int preconditioner = -1;
-	double xmax = 1.;
-	double ymax = 1.;
-	double zmax = 1.;
-	bool parallel = true;
-};
+#include "cas_cible_1.hxx"
 
 // legacy
-TestParameters parseCommandLineOptions(int& argc, char* argv[]) {
-  START_TIMER("parse_command_line_options");
-  TestParameters p;
+TestParameters parseCommandLineOptions_v2(int& argc, char* argv[]) {
+	START_TIMER("parse_command_line_options");
+	TestParameters p;
 
-  // options treatment
-  mfem::OptionsParser args(argc, argv);
-  args.AddOption(&p.mesh_file, "-m", "--mesh", "Mesh file to use.");
-  args.AddOption(&p.library, "-l", "--library", "Material library.");
-  args.AddOption(&p.order, "-o", "--order",
-                 "Finite element order (polynomial degree).");
-  
-  args.AddOption(&p.tcase, "-t", "--test-case",
-                 "identifier of the case : Exx->0, Eyy->1, Ezz->2, Exy->3, "
-                 "Exz->4, Eyz->5");
-  
-  args.AddOption(
-      &p.linearsolver, "-ls", "--linearsolver",
-      "identifier of the linear solver: 0 -> Default, 1 -> HypreFGMRES, 2 -> HypreGMRES, 3-> HyprePCG, 4 -> MUMPSSolver, 5 -> GMRESSolver, 6 -> CGSolver, 7 -> BiCGSTABSolver, 8 -> UMFPackSolver, 9 ->	MINRESSolver, 10 -> SLISolver");
+	// options treatment
+	mfem::OptionsParser args(argc, argv);
+	args.AddOption(&p.mesh_file, "-m", "--mesh", "Mesh file to use.");
+	args.AddOption(&p.library, "-l", "--library", "Material library.");
+	args.AddOption(&p.order, "-o", "--order",
+			"Finite element order (polynomial degree).");
 
-  args.AddOption(
-      &p.preconditioner, "-lp", "--preconditioner",
-      " 0 -> HypreDiagScale, 1 -> HypreParaSails, 2 -> HypreEuclid, 3 -> HypreBoomerAMG, 4 -> HypreILU, 5 -> ANY");
-  
-  args.Parse();
-  if (!args.Good()) {
-    if (mfem_mgis::getMPIrank() == 0)
-      args.PrintUsage(std::cout);
-    mfem_mgis::finalize();
-    exit(0);
-  }
-  if (p.mesh_file == nullptr) {
-    if (mfem_mgis::getMPIrank() == 0)
-      std::cout << "ERROR: Mesh file missing" << std::endl;
-    args.PrintUsage(std::cout);
-    mfem_mgis::abort(EXIT_FAILURE);
-  }
-  if (mfem_mgis::getMPIrank() == 0)
-    args.PrintOptions(std::cout);
-  if ((p.tcase < 0) || (p.tcase > 5)) {
-    std::cerr << "Invalid test case\n";
-    mfem_mgis::abort(EXIT_FAILURE);
-  }
-  return p;
+	args.AddOption(&p.tcase, "-t", "--test-case",
+			"identifier of the case : Exx->0, Eyy->1, Ezz->2, Exy->3, "
+			"Exz->4, Eyz->5");
+
+	args.AddOption(
+			&p.linearsolver, "-ls", "--linearsolver",
+			"identifier of the linear solver: 0 -> Default, 1 -> HypreFGMRES, 2 -> HypreGMRES, 3-> HyprePCG, 4 -> MUMPSSolver, 5 -> GMRESSolver, 6 -> CGSolver, 7 -> BiCGSTABSolver, 8 -> UMFPackSolver, 9 ->	MINRESSolver, 10 -> SLISolver");
+
+	args.AddOption(
+			&p.preconditioner, "-lp", "--preconditioner",
+			" 0 -> HypreDiagScale, 1 -> HypreParaSails, 2 -> HypreEuclid, 3 -> HypreBoomerAMG, 4 -> HypreILU, 5 -> ANY");
+
+	args.Parse();
+	if (!args.Good()) {
+		if (mfem_mgis::getMPIrank() == 0)
+			args.PrintUsage(std::cout);
+		mfem_mgis::finalize();
+		exit(0);
+	}
+	if (p.mesh_file == nullptr) {
+		if (mfem_mgis::getMPIrank() == 0)
+			std::cout << "ERROR: Mesh file missing" << std::endl;
+		args.PrintUsage(std::cout);
+		mfem_mgis::abort(EXIT_FAILURE);
+	}
+	if (mfem_mgis::getMPIrank() == 0)
+		args.PrintOptions(std::cout);
+	if ((p.tcase < 0) || (p.tcase > 5)) {
+		std::cerr << "Invalid test case\n";
+		mfem_mgis::abort(EXIT_FAILURE);
+	}
+	return p;
 }
 
 int execute(const TestParameters& p, const bool use_post_processing, const solver_name a_solv, const precond_name a_precond) {
 	constexpr const auto dim = mfem_mgis::size_type{3};
-	
+
 	//std::string string_solver = std::to_string(solv);
 	std::string string_solver 	= getName(a_solv);
 	std::string string_precond 	= getName(a_precond);
@@ -122,12 +104,12 @@ int execute(const TestParameters& p, const bool use_post_processing, const solve
 
 	// creating the finite element workspace
 	auto fed = std::make_shared<mfem_mgis::FiniteElementDiscretization>(
-	mfem_mgis::Parameters{{"MeshFileName", p.mesh_file},
-			    {"FiniteElementFamily", "H1"},
-			    {"FiniteElementOrder", p.order},
-			    {"UnknownsSize", dim},
-			    {"NumberOfUniformRefinements", p.parallel ? 0 : 0},
-			    {"Parallel", p.parallel}});
+			mfem_mgis::Parameters{{"MeshFileName", p.mesh_file},
+			{"FiniteElementFamily", "H1"},
+			{"FiniteElementOrder", p.order},
+			{"UnknownsSize", dim},
+			{"NumberOfUniformRefinements", p.parallel ? 0 : 0},
+			{"Parallel", p.parallel}});
 
 	mfem_mgis::PeriodicNonLinearEvolutionProblem problem(fed);
 
@@ -167,7 +149,7 @@ int execute(const TestParameters& p, const bool use_post_processing, const solve
 		}
 		problem.setMacroscopicGradientsEvolution([e](const double) { return e; });
 	} // end timer set_mgis_stuff
-	
+
 
 	setLinearSolver(problem, a_solv, a_precond);
 	setSolverParameters(problem);
@@ -176,9 +158,9 @@ int execute(const TestParameters& p, const bool use_post_processing, const solve
 	{
 		START_TIMER("add_postprocessing_and_outputs");
 		problem.addPostProcessing(
-			"ParaviewExportResults",
-			{{"OutputFileName", "PeriodicTestOutput-" + std::to_string(p.tcase)}}	
-		);
+				"ParaviewExportResults",
+				{{"OutputFileName", "PeriodicTestOutput-" + std::to_string(p.tcase)}}	
+				);
 	} // end timer add_postprocessing_and_outputs
 
 	mfem_mgis::NonLinearResolutionOutput solverStatistics;
@@ -189,14 +171,14 @@ int execute(const TestParameters& p, const bool use_post_processing, const solve
 		// solving the problem
 		profiling::output::printMessage("Solving ...");
 		measure = profiling::timers::chrono_section( [&](){
-			solverStatistics = problem.solve(0, 1);
-		});
+				solverStatistics = problem.solve(0, 1);
+				});
 		profiling::output::printMessage("... end solving");
 
 		// check status
 		if (!solverStatistics) {
 			profiling::output::printMessage("INFO: ", string_solver,"+",string_precond," FAILED");
-//			mfem_mgis::abort(EXIT_FAILURE);
+			//			mfem_mgis::abort(EXIT_FAILURE);
 		}
 	} // end timer Solve
 
@@ -210,10 +192,10 @@ int execute(const TestParameters& p, const bool use_post_processing, const solve
 		START_TIMER("get_statistics");
 		measure = profiling::output::reduce_max(measure);
 		profiling::output::printMessage(" solver: "	, getName(a_solv), 
-					" preconditioner: "	, getName(a_precond), 
-					" ite: "		, solverStatistics.iterations, 
-					" res: "		, solverStatistics.final_residual_norm, 
-					" time: "		, measure);
+				" preconditioner: "	, getName(a_precond), 
+				" ite: "		, solverStatistics.iterations, 
+				" res: "		, solverStatistics.final_residual_norm, 
+				" time: "		, measure);
 	} // end timer get_statistics
 	return(EXIT_SUCCESS);
 }
@@ -227,11 +209,11 @@ int main(int argc, char* argv[])
 	profiling::timers::init_timers();
 	constexpr bool use_post_processing = false;
 
-	auto p = parseCommandLineOptions(argc, argv);
+	auto p = parseCommandLineOptions_v2(argc, argv);
 	auto solver = solver_name(p.linearsolver);
 	auto preconditioner = precond_name(p.preconditioner);
 	int res;
-	if(match(solver,preconditioner))
+	if(cas_cible_1::match(solver,preconditioner))
 	{
 		res = execute(p, 
 				use_post_processing,
