+++ ex6/timer.cpp
@@ -1,6 +1,6 @@
 #include<timer.hpp>
 #ifdef __MPI
-	#include"mpi.h"
+#include"mpi.h"
 #endif
 #include<numeric>
 
@@ -27,7 +27,7 @@ namespace profiling
 			int size = -1;
 
 			double global = 0.0;
-                        MPI_Reduce(&a_duration, &global, 1, MPI_DOUBLE, MPI_MAX, master, MPI_COMM_WORLD); // master rank is 0
+			MPI_Reduce(&a_duration, &global, 1, MPI_DOUBLE, MPI_MAX, master, MPI_COMM_WORLD); // master rank is 0
 			return global;
 #else
 			return a_duration;
@@ -47,7 +47,7 @@ namespace profiling
 			m_iteration 	= 1;
 			m_level 	= 0;
 			m_mother 	= nullptr;
-			
+
 		}
 
 		profilingTimer::profilingTimer(std::string name, profilingTimer* mother): m_daughter(), m_duration(0)
@@ -59,230 +59,230 @@ namespace profiling
 		}
 
 		profilingTimer* 
-		profilingTimer::find(std::string name)
-		{
-			assert(this != nullptr);
-			for(auto it = m_daughter.begin() ; it < m_daughter.end() ; it++)
+			profilingTimer::find(std::string name)
 			{
-				if((*it)->m_name == name)
+				assert(this != nullptr);
+				for(auto it = m_daughter.begin() ; it < m_daughter.end() ; it++)
 				{
-					(*it)->m_iteration++;
-					return (*it);
+					if((*it)->m_name == name)
+					{
+						(*it)->m_iteration++;
+						return (*it);
+					}
+				}
+				profilingTimer* myTmp = new profilingTimer(name, this);
+				m_daughter.push_back(myTmp);
+				return myTmp;
+			}
+
+		void 
+			profilingTimer::printReplicate(size_t begin, size_t end, std::string motif)
+			{
+				for(size_t i = begin ; i < end ; i++) std::cout << motif;
+			}
+
+
+		void 
+			profilingTimer::space()
+			{
+				std::cout << " "; 
+			}
+
+		void 
+			profilingTimer::column()
+			{
+				std::cout << "|"; 
+			}
+
+		void 
+			profilingTimer::endline()
+			{
+				std::cout << std::endl; 
+			}
+
+		void 
+			profilingTimer::printBanner(size_t shift)
+			{
+				if(m_name == "root")
+				{
+#ifndef __MPI
+					profiling::output::printMessage(" MPI feature is disable for timers, if you use MPI please add -D__MPI ");
+#else
+					if(profiling::output::is_master()) {
+						profiling::output::printMessage(" MPI feature activated, rank 0:");
+#endif
+						std::string start_name = " |-- start timetable "; 
+						std::cout << start_name;
+						size_t end = shift+ nColumns*(cWidth+1) + 1;	
+						printReplicate(start_name.size(), end,"-");
+						column(); endline();
+						std::string name = " |    name";
+						std::cout << name;
+						printReplicate(name.size(),shift + 1," ");
+						for(size_t i =  0 ; i < nColumns ; i++)
+						{
+							column();
+							int size = cName[i].size();
+							printReplicate(0,(int(cWidth)-size - 1), " ");
+							std::cout << cName[i];
+							space();
+						}
+						column(); endline();
+						space(); column();
+						printReplicate(2, end,"-");
+						column(); endline();
+#ifdef __MPI
+					}
+#endif
 				}
 			}
-			profilingTimer* myTmp = new profilingTimer(name, this);
-			m_daughter.push_back(myTmp);
-			return myTmp;
-		}
-	 
-		void 
-		profilingTimer::printReplicate(size_t begin, size_t end, std::string motif)
-		{
-			for(size_t i = begin ; i < end ; i++) std::cout << motif;
-		}
-
 
 		void 
-		profilingTimer::space()
-		{
-			std::cout << " "; 
-		}
-
-		void 
-		profilingTimer::column()
-		{
-			std::cout << "|"; 
-		}
-
-		void 
-		profilingTimer::endline()
-		{
-			std::cout << std::endl; 
-		}
-
-		void 
-		profilingTimer::printBanner(size_t shift)
-		{
-			if(m_name == "root")
+			profilingTimer::printEnding(size_t shift)
 			{
-#ifndef __MPI
-				profiling::output::printMessage(" MPI feature is disable for timers, if you use MPI please add -D__MPI ");
+				if(m_name == "root")
+				{
+					if(profiling::output::is_master()) 
+					{
+						shift+= nColumns*(cWidth+1) + 1; // +1 for "|";
+						std::string end_name = " |-- end timetable " ;
+						std::cout << end_name;
+						printReplicate(end_name.size(),shift,"-");
+						column(); endline();
+					}
+				}
+			}
+
+		duration* 
+			profilingTimer::get_ptr_duration()
+			{
+				return &m_duration;
+			}
+
+		void 
+			profilingTimer::print(size_t shift, double total_time)
+			{
+				assert(total_time >= 0);
+				std::string cValue[nColumns];
+				if(profiling::output::is_master()) 
+				{
+					size_t realShift = shift;
+					space(); column(); space();
+					size_t currentShift = 3;
+					for(int i = 0 ; i < int(m_level) - 1; i++) 
+					{
+						int spaceSize = 3;
+						for(int j = 0 ; j < spaceSize ; j++) space();
+						currentShift += spaceSize;
+					}
+					if(m_level>0) {
+						std::cout << "|--";
+						currentShift += 3;
+					}
+					std::cout << "> "<< m_name;
+					currentShift += m_name.size() + 1;
+					printReplicate(currentShift, realShift, " ");
+
+					cValue[0] = std::to_string(m_iteration);	
+				}
+#ifdef __MPI
+				double local  = m_duration.count();	
+				int size = -1;
+				MPI_Comm_size(MPI_COMM_WORLD, &size);
+
+				assert(size > 0);
+				std::vector<double> list;
+
+				if(profiling::output::is_master()) list.resize(size);
+
+				MPI_Gather(&local,1,MPI_DOUBLE, list.data(), 1, MPI_DOUBLE, 0, MPI_COMM_WORLD); // master rank is 0
+
+
+				if(profiling::output::is_master())
+				{
+					const auto [min,max]	= std::minmax_element(list.begin(), list.end());
+					auto global_max 	= *max;
+					auto global_min 	= *min; 
+					auto sum 		= std::accumulate(list.begin(), list.end(), double(0.));
+					auto global_mean 	= sum / double(size);
+					auto part_time 		= (global_max / total_time ) * 100;
+
+					assert(global_mean >= 0);
+					assert(global_min >= 0);
+					assert(global_max >= 0);
+
+					assert(global_max >= global_mean);
+					assert(global_mean >= global_min);
+
+					cValue[1] = std::to_string( global_min);	
+					cValue[2] = std::to_string( global_mean);	
+					cValue[3] = std::to_string( global_max);	
+					cValue[4] = std::to_string( part_time) + "%";	
+					cValue[5] = std::to_string( (global_max/global_mean)-1) + "%";
+				}
 #else
-				if(profiling::output::is_master()) {
-					profiling::output::printMessage(" MPI feature activated, rank 0:");
+				cValue[1] = std::to_string( m_duration.count());	
+				cValue[2] = std::to_string( (m_duration.count()/total_time)*100 );	
 #endif
-					std::string start_name = " |-- start timetable "; 
-					std::cout << start_name;
-				        size_t end = shift+ nColumns*(cWidth+1) + 1;	
-					printReplicate(start_name.size(), end,"-");
-					column(); endline();
-					std::string name = " |    name";
-					std::cout << name;
-					printReplicate(name.size(),shift + 1," ");
+				if(profiling::output::is_master())
+				{
 					for(size_t i =  0 ; i < nColumns ; i++)
 					{
 						column();
-						int size = cName[i].size();
+						int size = cValue[i].size();
 						printReplicate(0,(int(cWidth)-size - 1), " ");
-						std::cout << cName[i];
+						std::cout << cValue[i];
 						space();
 					}
-					column(); endline();
-					space(); column();
-					printReplicate(2, end,"-");
-					column(); endline();
-#ifdef __MPI
-				}
-#endif
-			}
-		}
-
-		void 
-		profilingTimer::printEnding(size_t shift)
-		{
-			if(m_name == "root")
-			{
-				if(profiling::output::is_master()) 
-				{
-					shift+= nColumns*(cWidth+1) + 1; // +1 for "|";
-					std::string end_name = " |-- end timetable " ;
-					std::cout << end_name;
-					printReplicate(end_name.size(),shift,"-");
-					column(); endline();
+					column();endline();
 				}
 			}
-		}
-
-		duration* 
-		profilingTimer::get_ptr_duration()
-		{
-			return &m_duration;
-		}
-
-		void 
-		profilingTimer::print(size_t shift, double total_time)
-		{
-			assert(total_time >= 0);
-			std::string cValue[nColumns];
-			if(profiling::output::is_master()) 
-			{
-				size_t realShift = shift;
-				space(); column(); space();
-				size_t currentShift = 3;
-				for(int i = 0 ; i < int(m_level) - 1; i++) 
-				{
-					int spaceSize = 3;
-					for(int j = 0 ; j < spaceSize ; j++) space();
-					currentShift += spaceSize;
-				}
-				if(m_level>0) {
-					std::cout << "|--";
-					currentShift += 3;
-				}
-				std::cout << "> "<< m_name;
-				currentShift += m_name.size() + 1;
-				printReplicate(currentShift, realShift, " ");
-
-				cValue[0] = std::to_string(m_iteration);	
-			}
-#ifdef __MPI
-			double local  = m_duration.count();	
-			int size = -1;
-			MPI_Comm_size(MPI_COMM_WORLD, &size);
-
-			assert(size > 0);
-			std::vector<double> list;
-
-			if(profiling::output::is_master()) list.resize(size);
-
-			MPI_Gather(&local,1,MPI_DOUBLE, list.data(), 1, MPI_DOUBLE, 0, MPI_COMM_WORLD); // master rank is 0
-			
-
-			if(profiling::output::is_master())
-			{
-				const auto [min,max]	= std::minmax_element(list.begin(), list.end());
-				auto global_max 	= *max;
-				auto global_min 	= *min; 
-				auto sum 		= std::accumulate(list.begin(), list.end(), double(0.));
-				auto global_mean 	= sum / double(size);
-				auto part_time 		= (global_max / total_time ) * 100;
-
-				assert(global_mean >= 0);
-				assert(global_min >= 0);
-				assert(global_max >= 0);
-
-				assert(global_max >= global_mean);
-				assert(global_mean >= global_min);
-
-				cValue[1] = std::to_string( global_min);	
-				cValue[2] = std::to_string( global_mean);	
-				cValue[3] = std::to_string( global_max);	
-				cValue[4] = std::to_string( part_time) + "%";	
-				cValue[5] = std::to_string( (global_max/global_mean)-1) + "%";
-			}
-#else
-			cValue[1] = std::to_string( m_duration.count());	
-			cValue[2] = std::to_string( (m_duration.count()/total_time)*100 );	
-#endif
-			if(profiling::output::is_master())
-			{
-				for(size_t i =  0 ; i < nColumns ; i++)
-				{
-					column();
-					int size = cValue[i].size();
-					printReplicate(0,(int(cWidth)-size - 1), " ");
-					std::cout << cValue[i];
-					space();
-				}
-				column();endline();
-			}
-		}
 
 		std::string 
-		profilingTimer::getName()
-		{
-			return m_name;
-		}
+			profilingTimer::getName()
+			{
+				return m_name;
+			}
 
 		double 
-		profilingTimer::get_duration()
-		{
-			return m_duration.count();
-		}
+			profilingTimer::get_duration()
+			{
+				return m_duration.count();
+			}
 
 		std::size_t 
-		profilingTimer::get_iteration()
-		{
-			return m_iteration;
-		}
+			profilingTimer::get_iteration()
+			{
+				return m_iteration;
+			}
 
 		std::size_t 
-		profilingTimer::get_level()
-		{
-			return m_level;
-		}
-		
+			profilingTimer::get_level()
+			{
+				return m_level;
+			}
+
 		std::vector<profilingTimer*>& 
-		profilingTimer::get_daughter()
-		{
-			return m_daughter;
-		}
+			profilingTimer::get_daughter()
+			{
+				return m_daughter;
+			}
 
 		profilingTimer* 
-		profilingTimer::get_mother()
-		{
-			return m_mother;
-		}
+			profilingTimer::get_mother()
+			{
+				return m_mother;
+			}
 
 		void init_timers()
 		{
 			profiling::output::printMessage(" Init timers ");
 			profilingTimer*& root_timer_ptr 	= profiling::timers::get_timer<ROOT>() ;
-		        assert(root_timer_ptr == nullptr);	
+			assert(root_timer_ptr == nullptr);	
 			root_timer_ptr 			= new profilingTimer(); 
 			profilingTimer*& current 	= profiling::timers::get_timer<CURRENT>(); 
 			current 			= root_timer_ptr;
-		        assert(current != nullptr);	
+			assert(current != nullptr);	
 			profiling::timer::start_global_timer<ROOT>();
 		}
 
@@ -305,19 +305,19 @@ namespace profiling
 		}
 
 		inline 
-		void Timer::start()
-		{
-			m_start = steady_clock::now();
-		}
+			void Timer::start()
+			{
+				m_start = steady_clock::now();
+			}
 
 		inline 
-		void Timer::end()
-		{
-			assert(m_duration != nullptr && "duration has to be initialised");
-			m_stop = steady_clock::now();
-			*m_duration += m_stop - m_start;
-			assert(m_duration->count() >= 0);
-		}
+			void Timer::end()
+			{
+				assert(m_duration != nullptr && "duration has to be initialised");
+				m_stop = steady_clock::now();
+				*m_duration += m_stop - m_start;
+				assert(m_duration->count() >= 0);
+			}
 
 		Timer::~Timer() 
 		{
